Rename login route mixin to disambiguate
