+++ tests/unit/mixins/osf-token-login-route-test.js
@@ -1,12 +1,12 @@
 import Ember from 'ember';
-import OsfLoginRouteMixin from '../../../mixins/osf-login-route';
+import OsfTokenLoginRouteMixin from '../../../mixins/osf-token-login-route';
 import { module, test } from 'qunit';
 
-module('Unit | Mixin | osf login route');
+module('Unit | Mixin | osf token login route');
 
 // Replace this with your real tests.
 test('it works', function(assert) {
-  let OsfLoginRouteObject = Ember.Object.extend(OsfLoginRouteMixin);
+  let OsfLoginRouteObject = Ember.Object.extend(OsfTokenLoginRouteMixin);
   let subject = OsfLoginRouteObject.create();
   assert.ok(subject);
 });
